fix: stabilize tvOS reader input focus
Route tvOS DIVINA remote and keyboard events through the remote capture view so first-responder ownership stays consistent after the keyboard help prompt appears. Keep the prompt out of tvOS hit testing and share UIKit keyboard event mapping with the existing keyboard handler.
